move tests to a src dr
