+++ .github/workflows/Tests.yml
@@ -1,0 +1,31 @@
+name: Tests
+
+on:
+    push:
+        branches: ["main"]
+    pull_request:
+        branches: ["main"]
+
+jobs:
+    build_and_test:
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v6
+            - uses: actions/setup-node@v6
+            - name: Install
+              run: npm ci
+            - name: Test NodeJs
+              run: npm run test
+            - name: Test browsers
+              run: npx jackrabbit_webdriver .github/workflows/browsers.ubuntu.json tests/dist/mod.js
+    build_and_test_macos:
+        runs-on: macos-latest
+        steps:
+            - uses: actions/checkout@v6
+            - uses: actions/setup-node@v6
+            - name: Install
+              run: npm ci
+            - name: Test NodeJs
+              run: npm run test
+            - name: Test browsers
+              run: npx jackrabbit_webdriver .github/workflows/browsers.macos.json tests/dist/mod.js
