--- .github/workflows/build_and_test.yml
@@ -1,22 +1,0 @@
-name: Build and Test
-
-on:
-    push:
-        branches: ["main"]
-    pull_request:
-        branches: ["main"]
-
-jobs:
-    build_and_test:
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/checkout@v4
-            - uses: actions/setup-node@v4
-            - name: CI install
-              run: npm ci
-            - name: Format
-              run: npm run format
-            - name: Build tests
-              run: npm run build:tests
-            - name: Test
-              run: npm run test
