+++ .prettierignore
@@ -1,6 +1,6 @@
 node_modules/
 dist/
-tests/*.js
+tests/dist/
 
 LICENSE
 package-lock.json
