--- tests/assertion.js
@@ -1,16 +1,0 @@
-export function assert(expected, results) {
-    let [doc, error] = results;
-    if (error) {
-        return error;
-    }
-    if (expected !== doc) {
-        return `
-Expected:
-${expected}
-
-Found:
-${doc}
-
-`;
-    }
-}
